Reorganize the CLI test suite
