+++ pdm/cli/actions.py
@@ -127,10 +127,13 @@ def resolve_candidates_from_lockfile(
     return mapping
 
 
-def check_lockfile(project: Project) -> str | None:
+def check_lockfile(project: Project, raise_not_exist: bool = True) -> str | None:
     """Check if the lock file exists and is up to date. Return the update strategy."""
     if not project.lockfile_file.exists():
-        raise ProjectError("Lock file does not exist, nothing to install")
+        if raise_not_exist:
+            raise ProjectError("Lock file does not exist, nothing to install")
+        project.core.ui.echo("Lock file does not exist", fg="yellow", err=True)
+        return "all"
     elif not project.is_lockfile_compatible():
         project.core.ui.echo(
             "Lock file version is not compatible with PDM, installation may fail",
